ui调整，将一部分组件分离
